+++ firestore-bigquery-export/firestore-bigquery-change-tracker/src/__tests__/bigquery/e2e.test.ts
@@ -13,7 +13,7 @@ import { changeTracker, changeTrackerEvent } from "../fixtures/changeTracker";
 import { getBigQueryTableData } from "../fixtures/queries";
 import { firestore } from "firebase-admin";
 
-process.env.PROJECT_ID = "extensions-testing";
+process.env.PROJECT_ID = "dev-extensions-testing";
 
 const consoleLogSpy = jest.spyOn(logger, "log").mockImplementation();
 const consoleLogSpyWarn = jest.spyOn(logger, "warn").mockImplementation();
@@ -126,7 +126,7 @@ describe("e2e", () => {
           datasetId,
           tableId,
           timePartitioning: "DAY",
-          timePartitioningField: "created",
+          timePartitioningField: "timestamp",
           timePartitioningFieldType: "TIMESTAMP",
           timePartitioningFirestoreField: "created",
         }).record([event]);
@@ -141,7 +141,7 @@ describe("e2e", () => {
           tableId
         );
 
-        expect(changeLogRows[0].created.value).toBe(
+        expect(changeLogRows[0].timestamp.value).toBe(
           BigQuery.timestamp(created.toDate()).value
         );
       });