+++ firestore-bigquery-export/scripts/import/src/config.ts
@@ -171,10 +171,31 @@ const questions = [
     default: false,
   },
   {
+    message: "What's the URL of your transform function? (Optional)",
+    name: "transformFunctionUrl",
+    type: "input",
+    default: "",
+    validate: (value) => {
+      if (!value) return true;
+      try {
+        new URL(value);
+        return true;
+      } catch {
+        return "Please enter a valid URL or leave empty";
+      }
+    },
+  },
+  {
     message: "Would you like to use a local firestore emulator?",
     name: "useEmulator",
     type: "confirm",
     default: false,
+  },
+  {
+    message:
+      "Where would you like to output the failed imports? (Leave blank to skip)",
+    name: "failedBatchOutput",
+    type: "input",
   },
 ];
 
@@ -187,7 +208,7 @@ export async function parseConfig(): Promise<CliConfig | CliConfigError> {
       errors.push("Project is not specified.");
     }
     if (program.bigQueryProject === undefined) {
-      errors.push("BigQuery Project is not specified.");
+      program.bigQueryProject = program.project;
     }
     if (program.sourceCollectionPath === undefined) {
       errors.push("SourceCollectionPath is not specified.");
@@ -207,6 +228,15 @@ export async function parseConfig(): Promise<CliConfig | CliConfigError> {
     if (program.datasetLocation === undefined) {
       errors.push("DatasetLocation is not specified.");
     }
+
+    if (program.transformFunctionUrl) {
+      try {
+        new URL(program.transformFunctionUrl);
+      } catch {
+        errors.push("Transform function URL is invalid");
+      }
+    }
+
     if (!validateBatchSize(program.batchSize)) {
       errors.push("Invalid batch size.");
     }
@@ -239,6 +269,7 @@ export async function parseConfig(): Promise<CliConfig | CliConfigError> {
       useEmulator: program.useEmulator === "true",
       rawChangeLogName,
       cursorPositionFile,
+      failedBatchOutput: program.failedBatchOutput,
     };
   }
   const {
@@ -253,6 +284,7 @@ export async function parseConfig(): Promise<CliConfig | CliConfigError> {
     multiThreaded,
     useNewSnapshotQuerySyntax,
     useEmulator,
+    failedBatchOutput,
   } = await inquirer.prompt(questions);
 
   const rawChangeLogName = `${table}_raw_changelog`;
@@ -278,6 +310,7 @@ export async function parseConfig(): Promise<CliConfig | CliConfigError> {
     useEmulator: useEmulator,
     rawChangeLogName,
     cursorPositionFile,
+    failedBatchOutput,
   };
 }