+++ firestore-bigquery-export/CHANGELOG.md
@@ -4,6 +4,8 @@ feat - add basic materialized views support
 
 fix - do not add/update clustering if an invalid clustering field is present.
 
+docs - improve cross-project IAM documentation
+
 ## Version 0.1.56
 
 feat - improve sync strategy by immediately writing to BQ, and using cloud tasks only as a last resort