+++ firestore-bigquery-export/CHANGELOG.md
@@ -6,6 +6,8 @@ fix - do not add/update clustering if an invalid clustering field is present.
 
 docs - improve cross-project IAM documentation
 
+fix - emit correct events to extension, backwardly compatible.
+
 ## Version 0.1.56
 
 feat - improve sync strategy by immediately writing to BQ, and using cloud tasks only as a last resort